+++ _inc/client/recommendations/sidebar/product-card-upsell/style.scss
@@ -1,4 +1,4 @@
-@import '../../scss/variables/colors';
+@import '../../../scss/variables/colors';
 
 .jp-recommendations-product-card-upsell {
 	background: $white;
